% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nlmeYextra.R
\name{nlmeYextra}
\alias{nlmeYextra}
\title{Extra pdClasses and varClasses for special models using nlme}
\usage{
nlmeYextra()
}
\description{
nlmeYextra provides an new pdClass, `pdInd`, that models random effects that are independent of each other but not necessarily independent of the random intercepts.  The package also provides a variance class  variables whose
}

