% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pdInd.R
\name{pdFactor.pdInd}
\alias{pdFactor.pdInd}
\title{Factor for pdInd}
\usage{
pdFactor.pdInd(object)

pdFactor.pdInd(object)
}
\arguments{
\item{object}{a pdInd object representing a positive-definite variance matrix with covariances in the first row and column but zero covariances elsewhere.}

\item{object}{a 'pdInd' object from which the right-triangular factor of the variance matrix it represents will be extracted}
}
\value{
the columns of a lower triangular Choleski factor of the positive-definite matrix as a vector.

the full right-triangular factor, including zeros in the lower triangle, is returned as a vector in column order
}
\description{
Returns a factor for a 'left' log-Cholesky object for positive-definite variance matrix with zero covariances except in the first row and column. i.e.
$$
V = L'L
$$
with $L$ a lower-triangular matrix.

Function to compute the upper triangular factor of a pdInd object.
}
\examples{
mat <- pdInd(diag(1:4))
pdFactor(mat)
}

